fix bugs and improve test coverage found during package audit
- Fix smooth_chaikin() using seq.int() which returns c(1,0) for n=0,
  causing two loop iterations instead of zero when refinements=0
- Fix smooth_spline() rejecting integer vertex_factor (is.double -> is.numeric)
- Fix test_smooth.R checking polygon validity for line smoothing results
- Fix typos in source files and documentation (cannont, seqment, enpoints, etc.)
- Add tests for chaikin refinements=0, spline vertex_factor, densify
  max_distance on polygons, drop_crumbs drop_empty for lines

+++ R/smooth-spline.R
@@ -52,7 +52,7 @@ smooth_spline <- function(x, wrap = FALSE, vertex_factor = 5, n) {
   stopifnot(is_flag(wrap))
   if (missing(n)) {
     stopifnot(
-      is.double(vertex_factor),
+      is.numeric(vertex_factor),
       length(vertex_factor) == 1,
       vertex_factor >= 1
     )
